adding a simple gameobject system
there are lots of questions and possible issues with this
such as: how to handle default shaders and vertex arrays?

Look specifically at the SpriteComponent class, it initializes "defautls", which I'm not totally mad at, because that's a way to solve this problem I feel like, but also, it would probably be better to have the idea of a "Material" and then we have a "Default Sprite Material"

But also if you look at the header file you can see "defaultSpriteShader->SetMat4("u_Transform", transformMatrix") - are we sure that the shader will always have a transform called u_Transform? what if we cjhange it to "u_TransformMatrix"? - Having to set the transform uniform like this in the sprite shader is a bit weird!
Where else it can be done?